Safari-611.1.10/LayoutTests/media/W3C/video/canPlayType/canPlayType_codecs_order_3.html

<!doctype html>
<html>
 <head>
  <title>video.canPlayType() - codecs parameter order</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#mime-types">spec reference</a></p>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 var v = document.getElementById("v");
  assert_equals(
   v.canPlayType('video/mp4; codecs="avc1.42E01E, mp4a.40.2"'),
   v.canPlayType('video/mp4; codecs="mp4a.40.2, avc1.42E01E"'),
   "order of codecs parameters should have no effect");
});
  </script>
 </body>
</html>
