Safari-611.1.10/LayoutTests/platform/mac/media/W3C/video/canPlayType/canPlayType_codecs_order_3-expected.txt

video.canPlayType() - codecs parameter order

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is ""

TEST COMPLETE
spec reference
